SAAS-00: Update Feature Enum.
