+++ alembic/versions/104cc4c3e37b_update_enum_type.py
@@ -11,7 +11,7 @@ from alembic import op
 
 # revision identifiers, used by Alembic.
 revision: str = '104cc4c3e37b'
-down_revision: Union[str, None] = '3b27f6b61ac9'
+down_revision: Union[str, None] = None
 branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None
