+++ app/v2_0/application/service/employee_service.py
@@ -3,7 +3,7 @@ from sqlalchemy import select
 
 from app.v2_0.application.dto.dto_classes import ResponseDTO
 from app.v2_0.application.password_handler.reset_password import create_password_reset_code
-from app.v2_0.application.service.ucb_service import add_owner_to_ucb
+from app.v2_0.application.service.ucb_service import add_employee_to_ucb
 from app.v2_0.application.service.user_service import add_user_details
 from app.v2_0.application.utility.app_utility import check_if_company_and_branch_exist
 from app.v2_0.domain.models.branch_settings import BranchSettings
@@ -35,7 +35,7 @@ def set_employee_details(new_employee, branch_id, db):
 
 def assign_new_branch_to_existing_employee(employee, user, company_id, branch_id, db):
     """Adds the same employee to a different branch of the company"""
-    add_owner_to_ucb(employee, user, company_id, branch_id, db)
+    add_employee_to_ucb(employee, user, company_id, branch_id, db)
 
 
 def invite_employee(employee, user_id, company_id, branch_id, db):
@@ -54,7 +54,7 @@ def invite_employee(employee, user_id, company_id, branch_id, db):
                 new_employee = UsersAuth(user_email=employee.user_email, invited_by=inviter.user_email)
                 db.add(new_employee)
                 db.flush()
-                add_owner_to_ucb(employee, new_employee, company_id, branch_id, db)
+                add_employee_to_ucb(employee, new_employee, company_id, branch_id, db)
                 set_employee_details(new_employee, branch_id, db)
                 create_password_reset_code(employee.user_email, db)
                 db.commit()
