+++ app/v2_0/domain/models/enums.py
@@ -48,4 +48,3 @@ class Features(PyEnum):
     HR_APPLY_LEAVES = 0.6
     HR_MY_LEAVES = 0.7
     HR_TIMESHEET = 0.8
-    HR_FEATURE_1 = 0.9
